apps/platform/pyproject.toml
[project]
name = "rejeki_platform"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "jinja2",
    "python-dotenv",
    "itsdangerous",
    "bcrypt",
    "python-json-logger",
]

[project.scripts]
rejeki-platform = "rejeki_platform.server:main"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
